+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/client/LightTextureMixin.java
@@ -11,8 +11,8 @@ import org.spongepowered.asm.mixin.injection.At;
 public abstract class LightTextureMixin {
     @ModifyExpressionValue(method = "updateLightTexture", at = @At(value = "INVOKE", target = "Ljava/lang/Double;floatValue()F", ordinal = 1))
     public float onGamma(float value) {
-        if (Feature.isEnabled(Light.class) && Light.forceBrightness >= 0)
-            return (float) (Light.forceOnlyLowerBrightness ? Math.min(Light.forceBrightness, value) : Light.forceBrightness);
+        if (Feature.isEnabled(Light.class) && Light.forceDarkness >= 0)
+            return (float) (Light.forceOnlyLowerBrightness ? Math.min(Light.forceDarkness, value) : Light.forceDarkness);
         return value;
     }
 }